test(rstest): migrate cache-events config cases
