--- packages/webpack/cache-events-webpack-plugin/test/cases/cache-events/chunk-splitting/test.config.cjs
@@ -1,30 +1,0 @@
-/** @type {import("@lynx-js/test-tools").TConfigCaseConfig} */
-module.exports = {
-  bundlePath: [
-    'main.js',
-  ],
-  beforeExecute: () => {
-    global.lynxCoreInject = {
-      tt: {},
-    };
-    global.lynx = {
-      requireModuleAsync: (_request, callback) => {
-        callback(null, {
-          ids: ['0'],
-          modules: {
-            './cache-events/chunk-splitting/lib-common.js': function(
-              __unused_webpack___webpack_module__,
-              __webpack_exports__,
-              __webpack_require__,
-            ) {
-              const add = (a, b) => a + b;
-              __webpack_require__.d(__webpack_exports__, {
-                add: () => add,
-              });
-            },
-          },
-        });
-      },
-    };
-  },
-};
