--- packages/webpack/cache-events-webpack-plugin/test/cases/cache-events/not-splitting/lib-common.js
@@ -1,6 +1,0 @@
-/**
- * @param {number} a
- * @param {number} b
- * @returns {number}
- */
-export const add = (a, b) => a + b;
