--- packages/webpack/cache-events-webpack-plugin/test/cases/cache-events/not-splitting/test.config.cjs
@@ -1,6 +1,0 @@
-/** @type {import("@lynx-js/test-tools").TConfigCaseConfig} */
-module.exports = {
-  bundlePath: [
-    'main.js',
-  ],
-};
