--- packages/webpack/cache-events-webpack-plugin/test/cases/not-cache-events/lazy-bundle/index.js
@@ -1,7 +1,0 @@
-/// <reference types="@rspack/test-tools/rstest" />
-
-import('./lazy-component.js');
-
-it('should not have `__webpack_require__.lynx_ce`', () => {
-  expect(__webpack_require__['lynx_ce']).toBeFalsy();
-});
