--- packages/webpack/cache-events-webpack-plugin/test/cases/not-cache-events/lazy-bundle/lazy-component.js
@@ -1,3 +1,0 @@
-export default function LazyComponent() {
-  return null;
-}
