+++ packages/webpack/cache-events-webpack-plugin/test/config.test.ts
@@ -1,14 +1,12 @@
 // Copyright 2024 The Lynx Authors. All rights reserved.
 // Licensed under the Apache License Version 2.0 that can be found in the
 // LICENSE file in the root directory of this source tree.
-import path from 'node:path';
 import { fileURLToPath } from 'node:url';
 
-import { describeCases } from '@lynx-js/test-tools';
+import { createNormalCase, describeByWalk } from '@rspack/test-tools';
 
-const __dirname = path.dirname(fileURLToPath(import.meta.url));
+const __filename = fileURLToPath(import.meta.url);
 
-describeCases({
-  name: 'cache-events',
-  casePath: path.join(__dirname, 'cases'),
+describeByWalk(__filename, (name, src, dist) => {
+  createNormalCase(name, src, dist);
 });
