+++ packages/webpack/cache-events-webpack-plugin/test/configCases/cache-events/chunk-splitting/rspack.config.js
@@ -1,3 +1,6 @@
+// Copyright 2026 The Lynx Authors. All rights reserved.
+// Licensed under the Apache License Version 2.0 that can be found in the
+// LICENSE file in the root directory of this source tree.
 import { ChunkLoadingWebpackPlugin } from '@lynx-js/chunk-loading-webpack-plugin';
 
 import { LynxCacheEventsPlugin } from '../../../../lib/index.js';
@@ -25,23 +28,6 @@ export default {
   },
   plugins: [
     new ChunkLoadingWebpackPlugin(),
-    new LynxCacheEventsPlugin({
-      setupListTransformer: (setupList) => {
-        setupList.push(
-          `
-{
-  name: 'customCacheEvent',
-  setup: () => {
-    console.log('customCacheEvent setup');
-    return () => {
-      console.log('customCacheEvent teardown');
-    }
-  }
-}
-  `,
-        );
-        return setupList;
-      },
-    }),
+    new LynxCacheEventsPlugin(),
   ],
 };
