+++ packages/webpack/cache-events-webpack-plugin/test/configCases/cache-events/not-splitting/index.js
@@ -1,3 +1,6 @@
+// Copyright 2026 The Lynx Authors. All rights reserved.
+// Licensed under the Apache License Version 2.0 that can be found in the
+// LICENSE file in the root directory of this source tree.
 /// <reference types="@rspack/test-tools/rstest" />
 
 import { add } from './lib-common.js';
