+++ packages/webpack/cache-events-webpack-plugin/test/configCases/not-cache-events/lazy-bundle/lazy-component.js
@@ -1,0 +1,6 @@
+// Copyright 2026 The Lynx Authors. All rights reserved.
+// Licensed under the Apache License Version 2.0 that can be found in the
+// LICENSE file in the root directory of this source tree.
+export default function LazyComponent() {
+  return null;
+}
