+++ packages/webpack/cache-events-webpack-plugin/test/configCases/not-cache-events/lazy-bundle/rspack.config.js
@@ -1,3 +1,6 @@
+// Copyright 2026 The Lynx Authors. All rights reserved.
+// Licensed under the Apache License Version 2.0 that can be found in the
+// LICENSE file in the root directory of this source tree.
 import { ChunkLoadingWebpackPlugin } from '@lynx-js/chunk-loading-webpack-plugin';
 
 import { LynxCacheEventsPlugin } from '../../../../lib/index.js';
